% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamIndexWrapper.R
\docType{class}
\name{BamIndexWrapper}
\alias{BamIndexWrapper}
\alias{BamIndexWrapper-class}
\alias{stageObject,BamIndexWrapper-method}
\alias{loadBamIndexWrapper}
\title{Wrapper for a BAM index file}
\usage{
BamIndexWrapper(path)
}
\arguments{
\item{path}{String containing the path to a BAM index file.}
}
\value{
A BamIndexWrapper instance that can be used in \code{\link{stageObject}}.
}
\description{
Wrap a BAM index file for saving and loading in the \pkg{alabaster} framework.
}
\details{
The BamIndexWrapper class is a subclass of a \linkS4class{Wrapper},
so all of the methods of the latter can also be used here, e.g., \code{path}.
}
\examples{
# Mocking up a BAM index file.
test_bai <- system.file("extdata", "ex1.bam.bai", package="Rsamtools")

# Creating a BamIndexWrapper.
wrapped <- BamIndexWrapper(test_bai)
wrapped

# Staging the BamIndexWrapper.
dir <- tempfile()
library(alabaster.base)
info <- stageObject(wrapped, dir, "tab")
invisible(.writeMetadata(info, dir))
list.files(dir, recursive=TRUE)

# Loading it back again:
meta <- acquireMetadata(dir, "tab/file.bai")
loadObject(meta, dir)

}
\author{
Aaron Lun
}
