% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamWrapper.R
\docType{class}
\name{BamWrapper}
\alias{BamWrapper}
\alias{BamWrapper-class}
\alias{stageObject,BamWrapper-method}
\alias{loadBamWrapper}
\title{Wrapper for a BAM file}
\usage{
BamWrapper(path, index = NULL)
}
\arguments{
\item{path}{String containing the path to a BAM file.}

\item{index}{String specifying the path to an index file in tabix format, or \code{NULL} if no index is available.}
}
\value{
A BamWrapper instance that can be used in \code{\link{stageObject}}.
}
\description{
Wrap a BAM file for saving and loading in the \pkg{alabaster} framework.
}
\details{
The BamWrapper class is a subclass of a \linkS4class{IndexedWrapper},
so all of the methods of the latter can also be used here, e.g., \code{path}, \code{index}.

The \code{stageObject} method for BamWrapper classes will check the BAM file by scanning the header with \code{\link{scanBamHeader}}.
If an index is present, it will additionally run \code{\link{idxstatsBam}}.
}
\examples{
# Using a BAM file from Rsamtools.
fl <- system.file("extdata", "ex1.bam", package="Rsamtools", mustWork=TRUE)

# Creating a BamWrapper.
wrapped <- BamWrapper(fl)
wrapped

# Staging the BamWrapper.
dir <- tempfile()
library(alabaster.base)
info <- stageObject(wrapped, dir, "my_bam")
invisible(.writeMetadata(info, dir))
list.files(dir, recursive=TRUE)

# Loading it back again:
meta <- acquireMetadata(dir, "my_bam/file.bam")
loadObject(meta, dir)

}
\author{
Aaron Lun
}
