% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BedWrapper.R
\docType{class}
\name{BedWrapper}
\alias{BedWrapper}
\alias{loadBedWrapper}
\alias{BedWrapper-class}
\alias{show,BedWrapper-method}
\alias{stageObject,BedWrapper-method}
\title{Wrapper for a BED file}
\usage{
BedWrapper(path, compression = NULL, index = NULL)
}
\arguments{
\item{path}{String containing the path to a BED file.}

\item{compression}{String specifying the compression.
This should be one of \code{"none"}, \code{"gzip"}, \code{"bzip2"} or \code{"bgzip"}.
If \code{NULL}, this is inferred from the file's headers and suffix.}

\item{index}{String specifying the path to a tabix-formatted index file}
}
\value{
A BedWrapper instance that can be used in \code{\link{stageObject}}.
}
\description{
Wrap a BED file for saving and loading in the \pkg{alabaster} framework.
}
\examples{
# Mocking up a BED file.
tmp <- tempfile(fileext=".bed")
bed <- write("chr1\t2222\t33333", file=tmp)

# Creating a BedWrapper.
wrapped <- BedWrapper(tmp)
wrapped

# Staging the BedWrapper.
dir <- tempfile()
library(alabaster.base)
info <- stageObject(wrapped, dir, "my_bed")
invisible(.writeMetadata(info, dir))
list.files(dir, recursive=TRUE)

# Loading it back again:
meta <- acquireMetadata(dir, "my_bed/file.bed")
loadObject(meta, dir)

}
\author{
Aaron Lun
}
