% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigBedWrapper.R
\docType{class}
\name{BigBedWrapper}
\alias{BigBedWrapper}
\alias{BigBedWrapper-class}
\alias{stageObject,BigBedWrapper-method}
\alias{loadBigBedWrapper}
\title{Wrapper for a bigBed file}
\usage{
BigBedWrapper(path)
}
\arguments{
\item{path}{String containing the path to a bigBed file.}
}
\value{
A BigBedWrapper instance that can be used in \code{\link{stageObject}}.
}
\description{
Wrap a bigBed file for saving and loading in the \pkg{alabaster} framework.
}
\details{
The BigBedWrapper class is a subclass of a \linkS4class{Wrapper},
so all of the methods of the latter can also be used here, e.g., \code{path}.

The \code{stageObject} method for BigBedWrapper classes will check the bigBed file by searching for the bigBed magic number,
i.e., 0x8789F2EB or its byte-reversed form.
}
\examples{
# Mocking up a bigBed file.
test_path <- system.file("tests", package = "rtracklayer")
test_bb <- file.path(test_path, "test.bb")

# Creating a BigBedWrapper.
wrapped <- BigBedWrapper(test_bb)
wrapped

# Staging the BigBedWrapper.
dir <- tempfile()
library(alabaster.base)
info <- stageObject(wrapped, dir, "my_bb")
invisible(.writeMetadata(info, dir))
list.files(dir, recursive=TRUE)

# Loading it back again:
meta <- acquireMetadata(dir, "my_bb/file.bb")
loadObject(meta, dir)

}
\author{
Aaron Lun
}
