% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigWigWrapper.R
\docType{class}
\name{BigWigWrapper}
\alias{BigWigWrapper}
\alias{BigWigWrapper-class}
\alias{stageObject,BigWigWrapper-method}
\alias{loadBigWigWrapper}
\title{Wrapper for a bigWig file}
\usage{
BigWigWrapper(path)
}
\arguments{
\item{path}{String containing the path to a bigWig file.}
}
\value{
A BigWigWrapper instance that can be used in \code{\link{stageObject}}.
}
\description{
Wrap a bigWig file for saving and loading in the \pkg{alabaster} framework.
}
\details{
The BigWigWrapper class is a subclass of a \linkS4class{Wrapper},
so all of the methods of the latter can also be used here, e.g., \code{path}.

The \code{stageObject} method for BigWigWrapper classes will check the bigWig file by searching for the bigWig magic number,
i.e., 0x888FFC26 or its byte-reversed form.
}
\examples{
# Mocking up a bigWig file.
test_path <- system.file("tests", package = "rtracklayer")
test_bw <- file.path(test_path, "test.bw")

# Creating a BigWigWrapper.
wrapped <- BigWigWrapper(test_bw)
wrapped

# Staging the BigWigWrapper.
dir <- tempfile()
library(alabaster.base)
info <- stageObject(wrapped, dir, "my_bw")
invisible(.writeMetadata(info, dir))
list.files(dir, recursive=TRUE)

# Loading it back again:
meta <- acquireMetadata(dir, "my_bw/file.bw")
loadObject(meta, dir)

}
\author{
Aaron Lun
}
