% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FaidxWrapper.R
\docType{class}
\name{FaidxWrapper}
\alias{FaidxWrapper}
\alias{FaidxWrapper-class}
\alias{stageObject,FaidxWrapper-method}
\alias{loadFaidxWrapper}
\title{Wrapper for a FASTA index file}
\usage{
FaidxWrapper(path)
}
\arguments{
\item{path}{String containing the path to a FASTA index file.}
}
\value{
A FaidxWrapper instance that can be used in \code{\link{stageObject}}.
}
\description{
Wrap a FASTA index file for saving and loading in the \pkg{alabaster} framework.
}
\details{
The FaidxWrapper class is a subclass of a \linkS4class{Wrapper},
so all of the methods of the latter can also be used here, e.g., \code{path}.
}
\examples{
# Mocking up a FASTA index file.
test_fai <- system.file("extdata", "ce2dict1.fa.fai", package="Rsamtools")

# Creating a FaidxWrapper.
wrapped <- FaidxWrapper(test_fai)
wrapped

# Staging the FaidxWrapper.
dir <- tempfile()
library(alabaster.base)
info <- stageObject(wrapped, dir, "tab")
invisible(.writeMetadata(info, dir))
list.files(dir, recursive=TRUE)

# Loading it back again:
meta <- acquireMetadata(dir, "tab/file.fai")
loadObject(meta, dir)

}
\author{
Aaron Lun
}
