% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastaWrapper.R
\docType{class}
\name{FastaWrapper}
\alias{FastaWrapper}
\alias{FastaWrapper-class}
\alias{stageObject,FastaWrapper-method}
\alias{loadFastaWrapper}
\title{Wrapper for a FASTA file}
\usage{
FastaWrapper(path, compression = NULL, index = NULL)
}
\arguments{
\item{path}{String containing the path to a FASTA file.}

\item{compression}{String specifying the compression.
This should be one of \code{"none"}, \code{"gzip"}, \code{"bzip2"} or \code{"bgzip"}.
If \code{NULL}, this is inferred from the file's headers and suffix.}

\item{index}{String specifying the path to an faidx file, or \code{NULL} if no index is available.
If an index is supplied, the file should be uncompressed or bgzip-compressed.}
}
\value{
A FastaWrapper instance that can be used in \code{\link{stageObject}}.
}
\description{
Wrap a FASTA file for saving and loading in the \pkg{alabaster} framework.
}
\details{
The FastaWrapper class is a subclass of a \linkS4class{CompressedIndexedWrapper},
so all of the methods of the latter can also be used here,
e.g., \code{path}, \code{index}, \code{compression}.
}
\examples{
# Mocking up a FASTA file.
tmp <- file.path(fileext=".fa")
write(">FOOBAR\nacgtacgt", tmp)

# Creating a FastaWrapper.
wrapped <- FastaWrapper(tmp)
wrapped

# Staging the FastaWrapper.
dir <- tempfile()
library(alabaster.base)
info <- stageObject(wrapped, dir, "seq")
invisible(.writeMetadata(info, dir))
list.files(dir, recursive=TRUE)

# Loading it back again:
meta <- acquireMetadata(dir, "seq/file.fa")
loadObject(meta, dir)

}
\author{
Aaron Lun
}
