% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastqWrapper.R
\docType{class}
\name{FastqWrapper}
\alias{FastqWrapper}
\alias{FastqWrapper-class}
\alias{stageObject,FastqWrapper-method}
\alias{loadFastqWrapper}
\title{Wrapper for a FASTQ file}
\usage{
FastqWrapper(
  path,
  encoding,
  sequence.type = "DNA",
  compression = NULL,
  index = NULL
)
}
\arguments{
\item{path}{String containing the path to a FASTQ file.}

\item{encoding}{String specifying the encoding of the quality strings.
This should be one of \code{"phred"}, \code{"solexa"} or \code{"illumina"}.}

\item{compression}{String specifying the compression.
This should be one of \code{"none"}, \code{"gzip"}, \code{"bzip2"} or \code{"bgzip"}.
If \code{NULL}, this is inferred from the file's headers and suffix.}

\item{index}{String specifying the path to an faidx file, or \code{NULL} if no index is available.
If an index is supplied, the file should be uncompressed or bgzip-compressed.}
}
\value{
A FastqWrapper instance that can be used in \code{\link{stageObject}}.
}
\description{
Wrap a FASTQ file for saving and loading in the \pkg{alabaster} framework.
}
\details{
The FastqWrapper class is a subclass of a \linkS4class{CompressedIndexedWrapper},
so all of the methods of the latter can also be used here, e.g., \code{path}, \code{index}, \code{compression}.

The \code{stageObject} method for FastqWrapper classes will check the FASTQ file by reading the first few lines 
and attempting to import it into an \linkS4class{XStringSet} object using the relevant \pkg{Biostrings} functions, e.g., \code{\link{readDNAStringSet}}.
}
\examples{
# Mocking up a FASTQ file.
tmp <- tempfile(fileext=".fq")
write("@FOOBAR\nacgtacgt\n+134987382", tmp)

# Creating a FastqWrapper.
wrapped <- FastqWrapper(tmp)
wrapped

# Staging the FastqWrapper.
dir <- tempfile()
library(alabaster.base)
info <- stageObject(wrapped, dir, "seq")
invisible(.writeMetadata(info, dir))
list.files(dir, recursive=TRUE)

# Loading it back again:
meta <- acquireMetadata(dir, "seq/file.fa")
loadObject(meta, dir)

}
\author{
Aaron Lun
}
