% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileReference.R
\name{FileReference}
\alias{FileReference}
\alias{FileReference-class}
\alias{path,FileReference-method}
\alias{[[,FileReference-method}
\alias{[[<-,FileReference-method}
\alias{$,FileReference-method}
\alias{$<-,FileReference-method}
\alias{show,FileReference-method}
\title{Virtual file reference class}
\description{
A virtual class for file reference objects.
This implements common methods for \code{\link{path}}, \code{\link{[[}} and \code{\link{$}}.
}
\author{
Aaron Lun
}
