% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GffWrapper.R
\docType{class}
\name{GffWrapper}
\alias{GffWrapper}
\alias{GffWrapper-class}
\alias{stageObject,GffWrapper-method}
\alias{loadGffWrapper}
\title{Wrapper for a GFF file}
\usage{
GffWrapper(path, compression = NULL, index = NULL, format = NULL)
}
\arguments{
\item{path}{String containing the path to a GFF file.}

\item{compression}{String specifying the compression.
This should be one of \code{"none"}, \code{"gzip"}, \code{"bzip2"} or \code{"bgzip"}.
If \code{NULL}, this is inferred from the file's headers and suffix.}

\item{index}{String specifying the path to an index file in tabix format, or \code{NULL} if no index is available.
If an index is supplied, the file should be bgzip-compressed.}

\item{format}{String specifying the format of the GFF file.
This should be one of \code{"GFF2"} (i.e., GTF) or \code{"GFF3"}.
If \code{NULL}, this is inferred from the file extension of \code{path}.}
}
\value{
A GffWrapper instance that can be used in \code{\link{stageObject}}.
}
\description{
Wrap a GFF2/3 file for saving and loading in the \pkg{alabaster} framework.
}
\details{
The GffWrapper class is a subclass of a \linkS4class{CompressedIndexedWrapper},
so all of the methods of the latter can also be used here,
e.g., \code{path}, \code{index}, \code{compression}.
}
\examples{
# Using rtracklayer's example GFF file.
test_path <- system.file("tests", package = "rtracklayer")
test_gff3 <- file.path(test_path, "genes.gff3")

# Creating a GffWrapper.
wrapped <- GffWrapper(test_gff3)
wrapped

# Staging the GffWrapper.
dir <- tempfile()
library(alabaster.base)
info <- stageObject(wrapped, dir, "my_gff")
invisible(.writeMetadata(info, dir))
list.files(dir, recursive=TRUE)

# Loading it back again:
meta <- acquireMetadata(dir, "my_gff/file.gff3")
loadObject(meta, dir)

}
\author{
Aaron Lun
}
