% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabixWrapper.R
\docType{class}
\name{TabixWrapper}
\alias{TabixWrapper}
\alias{TabixWrapper-class}
\alias{stageObject,TabixWrapper-method}
\alias{loadTabixWrapper}
\title{Wrapper for a Tabix file}
\usage{
TabixWrapper(path)
}
\arguments{
\item{path}{String containing the path to a Tabix file.}
}
\value{
A TabixWrapper instance that can be used in \code{\link{stageObject}}.
}
\description{
Wrap a Tabix file for saving and loading in the \pkg{alabaster} framework.
}
\details{
The TabixWrapper class is a subclass of a \linkS4class{Wrapper},
so all of the methods of the latter can also be used here, e.g., \code{path}.
}
\examples{
# Mocking up a Tabix file.
test_tbx <- system.file("extdata", "example.gtf.gz.tbi", package="Rsamtools")

# Creating a TabixWrapper.
wrapped <- TabixWrapper(test_tbx)
wrapped

# Staging the TabixWrapper.
dir <- tempfile()
library(alabaster.base)
info <- stageObject(wrapped, dir, "tab")
invisible(.writeMetadata(info, dir))
list.files(dir, recursive=TRUE)

# Loading it back again:
meta <- acquireMetadata(dir, "tab/file.tbi")
loadObject(meta, dir)

}
\author{
Aaron Lun
}
