% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VcfWrapper.R
\docType{class}
\name{VcfWrapper}
\alias{VcfWrapper}
\alias{VcfWrapper-class}
\alias{stageObject,VcfWrapper-method}
\alias{loadVcfWrapper}
\title{Wrapper for a GFF file}
\usage{
VcfWrapper(path, compression = NULL, index = NULL, header_only = FALSE)
}
\arguments{
\item{path}{String containing the path to a GFF file.}

\item{compression}{String specifying the compression.
This should be one of \code{"none"}, \code{"gzip"}, \code{"bzip2"} or \code{"bgzip"}.
If \code{NULL}, this is inferred from the file's headers and suffix.}

\item{index}{String specifying the path to an index file in tabix format, or \code{NULL} if no index is available.
If an index is supplied, the file should be bgzip-compressed.}

\item{header_only}{Logical scalar specifying whether this file just contains the VCF headers without any actual data.
Occasionally useful for applications to perform preflight requests before downloading the full dataset.}
}
\value{
A VcfWrapper instance that can be used in \code{\link{stageObject}}.
}
\description{
Wrap a GFF2/3 file for saving and loading in the \pkg{alabaster} framework.
}
\details{
The VcfWrapper class is a subclass of a \linkS4class{CompressedIndexedWrapper},
so all of the methods of the latter can also be used here, e.g., \code{path}, \code{index}, \code{compression}.

The \code{stageObject} method for VcfWrapper classes will check the GFF file by reading the first few lines 
and attempting to import it into a GRanges via \code{readVCF} from the \pkg{VariantAnnotation} package.
If an index file is supplied, it will attempt to use that index in \code{\link{headerTabix}}.
}
\examples{
# Using VariantAnnotations's example file.
fl <- system.file("extdata", "structural.vcf", package="VariantAnnotation")

# Creating a VcfWrapper.
wrapped <- VcfWrapper(fl)
wrapped

# Staging the VcfWrapper.
dir <- tempfile()
library(alabaster.base)
info <- stageObject(wrapped, dir, "my_vcf")
invisible(.writeMetadata(info, dir))
list.files(dir, recursive=TRUE)

# Loading it back again:
meta <- acquireMetadata(dir, "my_vcf/file.vcf")
loadObject(meta, dir)

}
\author{
Aaron Lun
}
