% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BedWrapper.R
\docType{class}
\name{BedWrapper}
\alias{BedWrapper}
\alias{BedWrapper-class}
\alias{stageObject,BedWrapper-method}
\alias{loadBedWrapper}
\title{Wrapper for a BED file}
\usage{
BedWrapper(path, compression = NULL, index = NULL)
}
\arguments{
\item{path}{String containing the path to a BED file.}

\item{compression}{String specifying the compression.
This should be one of \code{"none"}, \code{"gzip"}, \code{"bzip2"} or \code{"bgzip"}.
If \code{NULL}, this is inferred from the file's headers and suffix.}

\item{index}{String specifying the path to an index file in tabix format, or \code{NULL} if no index is available.
If an index is supplied, the file should be bgzip-compressed.}
}
\value{
A BedWrapper instance that can be used in \code{\link{stageObject}}.
}
\description{
Wrap a BED file for saving and loading in the \pkg{alabaster} framework.
}
\details{
The BedWrapper class is a subclass of a \linkS4class{CompressedIndexedWrapper},
so all of the methods of the latter can also be used here, e.g., \code{path}, \code{index}, \code{compression}.

The \code{stageObject} method for BedWrapper classes will check the BED file by reading the first few lines 
and attempting to import it into a GRanges via \code{\link{import.bed}} or \code{\link{import.bed15}}.
}
\examples{
# Mocking up a BED file.
tmp <- tempfile(fileext=".bed")
bed <- write("chr1\t2222\t33333", file=tmp)

# Creating a BedWrapper.
wrapped <- BedWrapper(tmp)
wrapped

# Staging the BedWrapper.
dir <- tempfile()
library(alabaster.base)
info <- stageObject(wrapped, dir, "my_bed")
invisible(.writeMetadata(info, dir))
list.files(dir, recursive=TRUE)

# Loading it back again:
meta <- acquireMetadata(dir, "my_bed/file.bed")
loadObject(meta, dir)

}
\author{
Aaron Lun
}
