% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmtWrapper.R
\docType{class}
\name{GmtWrapper}
\alias{GmtWrapper}
\alias{GmtWrapper-class}
\alias{stageObject,GmtWrapper-method}
\alias{show,GmtWrapper-method}
\alias{loadGmtWrapper}
\title{Wrapper for a GMT file}
\usage{
GmtWrapper(path, compression = NULL)
}
\arguments{
\item{path}{String containing the path to a GMT file.}

\item{compression}{String specifying the compression.
This should be one of \code{"none"}, \code{"gzip"}, \code{"bzip2"} or \code{"bgzip"}.
If \code{NULL}, this is inferred from the file's headers and suffix.}
}
\value{
A GmtWrapper instance that can be used in \code{\link{stageObject}}.
}
\description{
Wrap a GMT file for saving and loading in the \pkg{alabaster} framework.
}
\details{
The GmtWrapper class is a subclass of a \linkS4class{CompressedIndexedWrapper},
so all of the methods of the latter can also be used here,
e.g., \code{path}, \code{index}, \code{compression}.
}
\examples{
# Mocking up a GMT file.
tmp <- tempfile(fileext=".gmt")
write("SET1\tdescription\tgene1\tgene2\tgene3", file=tmp)

# Creating a GmtWrapper.
wrapped <- GmtWrapper(tmp)
wrapped

# Staging the GmtWrapper.
dir <- tempfile()
library(alabaster.base)
info <- stageObject(wrapped, dir, "my_gmt")
invisible(.writeMetadata(info, dir))
list.files(dir, recursive=TRUE)

# Loading it back again:
meta <- acquireMetadata(dir, "my_gmt/file.gmt")
loadObject(meta, dir)

}
\author{
Aaron Lun
}
